pyproject.toml
[project]
name = "rag-eval-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.6.0",
    "duckdb>=1.4.4",
    "faiss-cpu>=1.13.2",
    "google-cloud-bigquery>=3.40.1",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
